made the game run again after it finishes
Was unable to get the code running on my machine, since I am on Mac OS X, and windows.h appears to be a windows-specific header file, however, if my understanding of the code is correct, this should allow the game to be run again without re-running the program. However, I was unable to test and/or debug this.